rf: ZENKO-143 move attachReqUids() helper
Move attachReqUids() helper to lib/client/utils.js. This to done to
share the simple but useful logic across extensions.
