+++ lib/clients/utils.js
@@ -6,4 +6,6 @@ function attachReqUids(s3req, log) {
     });
 }
 
-module.exports = attachReqUids;
+module.exports = {
+    attachReqUids,
+};
